test:integration - fix renamed test data file
