+++ test/integration/lib/currency-localization.js
@@ -4,7 +4,7 @@ const {
   queryAsync,
   findAsync,
 } = require('../../lib/util')
-const fetchMockResponses = require('../../e2e/fetch-mocks.js')
+const fetchMockResponses = require('../../e2e/fetch-mocks.json')
 
 QUnit.module('currency localization')
